+++ NINA.Photon.Plugin.ASA/Properties/AssemblyInfo.cs
@@ -20,8 +20,8 @@ using System.Runtime.InteropServices;
 // [MANDATORY] The assembly versioning
 //Should be incremented for each new release build of a plugin
 
-[assembly: AssemblyVersion("3.2.8.16")]
-[assembly: AssemblyFileVersion("3.2.8.16")]
+[assembly: AssemblyVersion("3.2.8.17")]
+[assembly: AssemblyFileVersion("3.2.8.17")]
 
 // [MANDATORY] The name of your plugin
 [assembly: AssemblyTitle("ASA Tools")]